move tests and rename main directory
